libblueprint now reads Lua Code
The Lua code itself is saved in an array called 'BlockStringData'.
The array 'BlockStringDataIds' implicitely associates a random numerical ID
to each BlockStringData. (Per same index)
This numerical id is set as the 4th value in the per-block BP1 array.

(Confusing, isn't it?)

I didn't observe any semantic meaning to the numerical ID, so I won't
save it since I can simply generate it for serialization.